+++ api/index.js
@@ -352,12 +352,10 @@ var PrismaDB = class {
         return created;
       },
       delete: async (where) => {
-        const deleted = await prisma.habitGroupMember.delete({
+        const deleted = await prisma.habitGroupMember.deleteMany({
           where: {
-            group_id_user_id: {
-              group_id: where.group_id,
-              user_id: where.user_id
-            }
+            group_id: where.group_id,
+            user_id: where.user_id
           }
         });
         return deleted;